requirements.txt
python-binance
python-telegram-bot
ta
pandas
numpy
scikit-learn
python-dotenv
